Refactor Pytest fixtures for better manageability
