+++ tests/e2e/conftest.py
@@ -1,0 +1,1 @@
+"""Placeholder file for e2e fixtures."""
