+++ tests/fixtures/__init__.py
@@ -1,0 +1,10 @@
+"""
+Fixture modules for KP Analysis Toolkit tests.
+
+This package organizes test fixtures by domain while maintaining
+cross-domain accessibility through pytest's discovery mechanism.
+"""
+
+from tests.fixtures.di_fixtures import *  # noqa: F403
+from tests.fixtures.mock_fixtures import *  # noqa: F403
+from tests.fixtures.system_fixtures import *  # noqa: F403
