+++ tests/fixtures/di_fixtures/__init__.py
@@ -1,0 +1,6 @@
+"""Package to organize DI test fixtures for the KP Analysis Toolkit."""
+
+from tests.fixtures.di_fixtures.core_fixtures import *  # noqa: F403
+from tests.fixtures.di_fixtures.environment_fixtures import *  # noqa: F403
+from tests.fixtures.di_fixtures.service_fixtures import *  # noqa: F403
+# Future imports will go here as you add toolkit-specific fixtures
