+++ tests/performance/conftest.py
@@ -1,0 +1,1 @@
+"""Placeholder for performance fixtures."""
